modernized types

+++ day02/main.c
@@ -1,3 +1,5 @@
+#include <inttypes.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -12,12 +14,12 @@ int main(int argc, char **argv) {
     return 1;
   }
 
-  static long long ids[100000];
+  static int64_t ids[100000];
   int n = 0;
-  long long total = 0;
+  int64_t total = 0;
 
-  for (long long mult = 10; mult <= 100000; mult *= 10) {
-    for (long long chunk = mult / 10; chunk <= mult - 1; chunk++) {
+  for (int64_t mult = 10; mult <= 100000; mult *= 10) {
+    for (int64_t chunk = mult / 10; chunk <= mult - 1; chunk++) {
       ids[n++] = chunk * mult + chunk;
     }
   }
@@ -37,7 +39,7 @@ int main(int argc, char **argv) {
 
       char *endptr;
 
-      long long start = strtoll(p, &endptr, 10);
+      int64_t start = strtoll(p, &endptr, 10);
 
       // duck out if nothing more to look at
       if (p == endptr) {
@@ -51,7 +53,7 @@ int main(int argc, char **argv) {
       }
       p++;
 
-      long long end = strtoll(p, &endptr, 10);
+      int64_t end = strtoll(p, &endptr, 10);
       if (p == endptr)
         break;
       p = endptr;
@@ -66,7 +68,7 @@ int main(int argc, char **argv) {
       }
     }
   }
-  printf("%lld\n", total);
+  printf("%" PRId64 "\n", total);
   fclose(fp);
   return 0;
 }
